2/2 refactor, but now it needs propagation
