+++ helm/templates/deploy.py
@@ -32,7 +32,7 @@ import glob
 runfile_deps = [path for path in map(os.path.abspath, glob.glob('external/*/*'))]
 sys.path = runfile_deps + sys.path
 
-from common.cloudsmith.uploader import Uploader
+from common.uploader.uploader import Uploader
 
 if len(sys.argv) != 2:
     raise ValueError('Should pass only <snapshot|release> as arguments')
